Add controllers and services

+++ src/main/java/hr/fer/zavrsni1500/itshop/model/Product.java
@@ -16,7 +16,7 @@ public class Product {
 
     private String name;
 
-    private Float price;
+    private double price;
 
     private String description;
 }
